+++ api_gateway/app/route/computation.py
@@ -6,6 +6,7 @@ from celery.result import AsyncResult
 from climatoology.base.event import ComputationState
 from climatoology.utility.exception import ClimatoologyUserError, InputValidationError
 from fastapi import APIRouter, WebSocketException
+from fastapi_cache.decorator import cache
 from starlette.requests import Request
 from starlette.websockets import WebSocket
 
@@ -29,6 +30,7 @@ async def subscribe_compute_status(websocket: WebSocket, correlation_uuid: UUID)
     description='Get the state of the computation with messages. States are equal to the celery computation states '
     'described here: https://docs.celeryq.dev/en/stable/userguide/tasks.html#built-in-states',
 )
+@cache(expire=2)
 def get_computation_status(correlation_uuid: UUID, request: Request) -> ComputationStateInfo:
     computation_uuid = request.app.state.platform.backend_db.resolve_computation_id(
         correlation_uuid