+++ CHANGELOG.md
@@ -5,13 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://gitlab.heigit.org/climate-action/api-gateway/-/compare/2.1.0...main)
+## [Unreleased](https://gitlab.heigit.org/climate-action/api-gateway/-/compare/2.1.1...main)
+
+### Deprecated
+
+- the websocket endpoint is currently broken
+
+### Changed
+
+- update climatoology to version 6.0.0 that uses Celery as the underlying task management library
+- made `aoi` a separate input parameter for compute requests
+
+### Added
+
+- `state` endpoint for computations ([#12](https://gitlab.heigit.org/climate-action/api-gateway/-/issues/12))
 
 ## [2.1.1](https://gitlab.heigit.org/climate-action/api-gateway/-/releases/2.1.1) - 2024-09-17
 
 ### Changed
 
-- `fetch`-endpoints now redirect to a pre-signed S3 url instead of providing a file response [#11](https://gitlab.heigit.org/climate-action/api-gateway/-/issues/11)
+- `fetch`-endpoints now redirect to a pre-signed S3 url instead of providing a file
+  response [#11](https://gitlab.heigit.org/climate-action/api-gateway/-/issues/11)
 - update climatoology to 5.2.0
 
 ### Fixed
@@ -22,12 +36,14 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 ### Changed
 
-- swagger and redoc can now be disabled through a feature switch env-var `DISABLE_SWAGGER`. This is necessary to prevent external calls by the swagger website.
+- swagger and redoc can now be disabled through a feature switch env-var `DISABLE_SWAGGER`. This is necessary to prevent
+  external calls by the swagger website.
 - dependency cleanup
 
 ### Added
 
-- the openapi.json now exposes the api and the climatoology version automatically ([#8](https://gitlab.heigit.org/climate-action/api-gateway/-/issues/8))
+- the openapi.json now exposes the api and the climatoology version
+  automatically ([#8](https://gitlab.heigit.org/climate-action/api-gateway/-/issues/8))
 
 ## [2.0.0](https://gitlab.heigit.org/climate-action/api-gateway/-/releases/2.0.0) - 2024-06-10
 
@@ -38,7 +54,8 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 ### Fixed
 
 - the logical check for the existence of an artifact
-- an issue where websockets were blocking all resources in the api rendering it a single user application ([#3](https://gitlab.heigit.org/climate-action/api-gateway/-/issues/3))
+- an issue where websockets were blocking all resources in the api rendering it a single user
+  application ([#3](https://gitlab.heigit.org/climate-action/api-gateway/-/issues/3))
 
 ### Added
 
@@ -49,4 +66,5 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 ### Added
 
-- extracting the API gateway from [climatoology](https://gitlab.heigit.org/climate-action/climatoology) where it was embedded until now
+- extracting the API gateway from [climatoology](https://gitlab.heigit.org/climate-action/climatoology) where it was
+  embedded until now